requirements.txt
agno
azure-identity
python-dotenv
pypdf2
pydantic
typing-extensions
streamlit 